+++ src/GlobalUsings.cs
@@ -1,3 +1,4 @@
 ﻿global using System.Runtime.CompilerServices;
 
 global using Gasolutions.Core.Patterns.Result.Errors;
+global using Gasolutions.Core.Patterns.Result.Localization;